fix: address settlement review findings
