+++ src/application/RuntimeMutationGate.ts
@@ -4,6 +4,12 @@ type Release = () => void;
 export default class RuntimeMutationGate {
   #tail: Promise<void> = Promise.resolve();
 
+  /**
+   * Runs one authoritative mutation after every previously queued mutation.
+   *
+   * Operations must not call this method directly or transitively on the same
+   * gate: reentrant use waits on the active operation and deadlocks.
+   */
   async run<T>(operation: () => Promise<T>): Promise<T> {
     let release: Release | undefined;
     const turn = new Promise<void>((resolve) => {
