+++ test/unit/application/RuntimeSettlement.test.ts
@@ -1,4 +1,4 @@
-import { describe, expect, it, vi } from 'vitest';
+import { describe, expect, it } from 'vitest';
 
 import { settleRuntimePlan } from '../../../src/application/RuntimeSettlement.ts';
 import AdmissionEvaluation from '../../../src/domain/admission/AdmissionEvaluation.ts';
@@ -8,9 +8,7 @@ import PluralAdmission from '../../../src/domain/admission/PluralAdmission.ts';
 import PluralityWitness from '../../../src/domain/admission/PluralityWitness.ts';
 import type { AdmissionOutcome } from '../../../src/domain/api/AdmissionOutcome.ts';
 import { projectAdmissionOutcome } from '../../../src/domain/api/AdmissionOutcomeRuntime.ts';
-import type { SettlementSourceRuntime } from '../../../src/domain/api/LaneSettlementRuntime.ts';
-import { bindSettlementPlan } from '../../../src/domain/api/SettlementPlanRuntime.ts';
-import SettlementPlan from '../../../src/domain/settlement/SettlementPlan.ts';
+import { boundSettlementOutcomeHarness } from '../../helpers/SettlementHarness.ts';
 
 const EVALUATION = new AdmissionEvaluation({
   sourceParticipantId: 'lane:strand:draft',
@@ -28,71 +26,15 @@ describe('Runtime settlement outcome preservation', () => {
     ['plural', pluralOutcome()],
     ['conflict', conflictOutcome()],
   ] as const)('does not promote a %s plan', async (_kind, outcome) => {
-    const owner = Object.freeze({});
-    const plan = settlementPlan();
-    const snapshot = Object.freeze({
-      baseTargetFrontierRef: plan.targetFrontier.id,
-      frontierRef: plan.sourceFrontier.id,
-      proposalDigest: plan.proposalDigest,
-      status: 'ready' as const,
-      targetFrontierRef: plan.targetFrontier.id,
-    });
-    const promote = vi.fn(async () => Object.freeze({
-      accepted: true,
-      evidence: undefined,
-      reason: undefined,
-    }));
-    const digest = vi.fn(async (parts: readonly string[]) =>
-      parts[0] === 'law' ? plan.lawDigest : plan.policyDigest
-    );
-    const sourceRuntime: SettlementSourceRuntime = Object.freeze({
-      kind: 'source',
-      capture: async () => snapshot,
-      digest,
-      runExclusive: async (operation) =>
-        await operation(Object.freeze({
-          capture: async () => snapshot,
-          digest,
-          promote,
-        })),
-    });
-    const evidence = Object.freeze({
-      basis: plan.targetFrontier,
-      support: Object.freeze([
-        plan.sourceFrontier,
-        Object.freeze({ id: plan.proposalDigest }),
-      ]),
-    });
-    bindSettlementPlan(plan, {
-      evidence,
-      outcome,
-      owner,
-      source: Object.freeze({ kind: 'strand', name: 'draft' }),
-      sourceRuntime,
-      sourceSnapshot: snapshot,
-      target: Object.freeze({ kind: 'worldline', name: 'events' }),
-    });
+    const harness = boundSettlementOutcomeHarness(outcome);
 
-    const receipt = await settleRuntimePlan(plan, owner);
+    const receipt = await settleRuntimePlan(harness.plan, harness.owner);
 
     expect(receipt.outcome).toBe(outcome);
-    expect(receipt.evidence).toEqual(evidence);
-    expect(promote).not.toHaveBeenCalled();
+    expect(receipt.evidence).toEqual(harness.evidence);
+    expect(harness.promote).not.toHaveBeenCalled();
   });
 });
-
-function settlementPlan(): SettlementPlan {
-  return new SettlementPlan({
-    planDigest: 'plan:settlement',
-    sourceLaneId: 'lane:strand:draft',
-    targetLaneId: 'lane:worldline:events',
-    sourceFrontier: Object.freeze({ id: 'frontier:source' }),
-    targetFrontier: Object.freeze({ id: 'frontier:target' }),
-    proposalDigest: 'proposal:draft',
-    lawDigest: 'law:settlement',
-    policyDigest: 'policy:settlement',
-  });
-}
 
 function pluralOutcome(): AdmissionOutcome {
   return projectAdmissionOutcome(
